refactor(sdk): separate keys functions
