+++ packages/sdk/ts/maciKeys/__tests__/publicKeys.test.ts
