+++ packages/sdk/ts/maciKeys/index.ts
@@ -1,0 +1,3 @@
+export { generateKeypair } from "./keypair";
+export { generateMaciPublicKey } from "./publicKeys";
+export type { IGenerateKeypairArgs, IGenerateKeypairData } from "./types";
