+++ packages/sdk/ts/verifyingKeys/checkVerifyingKeys.ts
