+++ packages/sdk/ts/verifyingKeys/index.ts
@@ -1,6 +1,4 @@
 export { checkVerifyingKeys } from "./checkVerifyingKeys";
-export { generateKeypair } from "./keypair";
-export { generateMaciPublicKey } from "./publicKeys";
 export { setVerifyingKeys } from "./setVerifyingKeys";
 export { getAllOnChainVks, compareVks, extractAllVks, extractVkToFile } from "./utils";
 export type {
@@ -9,7 +7,5 @@ export type {
   IExtractAllVksArgs,
   IMaciVks,
   ISetVerifyingKeysArgs,
-  IGenerateKeypairArgs,
-  IGenerateKeypairData,
   ICheckVerifyingKeysArgs,
 } from "./types";
