+++ packages/sdk/ts/verifyingKeys/types.ts
@@ -37,31 +37,6 @@ export interface IGetAllVksArgs {
 }
 
 /**
- * Interface for the arguments for generate keypair command
- */
-export interface IGenerateKeypairArgs {
-  /**
-   * Seed value for keypair
-   */
-  seed?: bigint;
-}
-
-/**
- * Interface for the return data type for generate keypair command
- */
-export interface IGenerateKeypairData {
-  /**
-   * Serialized public key
-   */
-  publicKey: string;
-
-  /**
-   * Serialized private key
-   */
-  privateKey: string;
-}
-
-/**
  * MACI's verifying keys
  */
 export interface IMaciVerifyingKeys {
