+++ packages/sdk/ts/verifyingKeys/utils.ts
